requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
motor==3.5.0
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
google-auth-oauthlib==1.2.1
google-api-python-client==2.123.0
google-auth-httplib2==0.2.0
PyJWT==2.10.1
pywebpush==1.14.0
requests==2.32.3
python-multipart==0.0.6
aiofiles==24.1.0
google-generativeai==0.3.0
httpx==0.27.0
